[SYCL] move usm_allocator to UR
It will be shared by level zero plugin and UR adapters
