+++ sycl/plugins/level_zero/pi_level_zero.cpp
@@ -27,7 +27,6 @@
 #include <zet_api.h>
 
 #include "ur_bindings.hpp"
-#include "usm_allocator.hpp"
 
 extern "C" {
 // Forward declarartions.
